Finished form
Finished the form for the page which contains multiple inputs for the user to fill out, as well as multiple selection lists and options the user must fill out to complete their purchase. All that I need to do now is add the styles to the form and the elements within.